Finish converting BookBags to async
* Migrate create/delete to UserService and async
* Add ability to create list with description
