+++ Source/Evergreen/ActorService.swift
@@ -131,46 +131,6 @@ class ActorService {
         return req.gatewayOptionalObjectResponse()
     }
 
-    static func fetchBookBagContents(authtoken: String, bookBag: BookBag) -> Promise<Void> {
-        let query = "container(bre,bookbag,\(bookBag.id),\(authtoken))"
-        let options = ["limit": 999]
-        let req = Gateway.makeRequest(service: API.search, method: API.multiclassQuery, args: [options, query, 0], shouldCache: false)
-        let promise = req.gatewayObjectResponse().then { (obj: OSRFObject) -> Promise<(OSRFObject)> in
-            bookBag.initVisibleIds(fromQueryObj: obj)
-            let req2 = Gateway.makeRequest(service: API.actor, method: API.containerFlesh, args: [authtoken, API.containerClassBiblio, bookBag.id], shouldCache: false)
-            return req2.gatewayObjectResponse()
-        }.done { obj in
-            bookBag.loadItems(fromFleshedObj: obj)
-        }
-        return promise
-    }
-    
-    static func createBookBag(authtoken: String, userID: Int, name: String) -> Promise<Void> {
-        let obj = OSRFObject([
-            "btype": API.containerTypeBookbag,
-            "name": name,
-            "pub": false,
-            "owner": userID,
-        ], netClass: "cbreb")
-        let req = Gateway.makeRequest(service: API.actor, method: API.containerCreate, args: [authtoken, API.containerClassBiblio, obj], shouldCache: false)
-        let promise = req.gatewayResponse().done { resp in
-            if let str = resp.str {
-                os_log("[bookbag] createBag %@ result %@", name, str)
-            }
-        }
-        return promise
-    }
-    
-    static func deleteBookBag(authtoken: String, bookBagId: Int) -> Promise<Void> {
-        let req = Gateway.makeRequest(service: API.actor, method: API.containerDelete, args: [authtoken, API.containerClassBiblio, bookBagId], shouldCache: false)
-        let promise = req.gatewayResponse().done { resp in
-            if let str = resp.str {
-                os_log("[bookbag] bag %d deleteBag result %@", bookBagId, str)
-            }
-        }
-        return promise
-    }
-    
     static func addItemToBookBag(authtoken: String, bookBagId: Int, recordId: Int) -> Promise<Void> {
         let obj = OSRFObject([
             "bucket": bookBagId,
